refactored cards to be on separate component
